requirements.txt
langchain
langchain-community
unstructured
markdown
sentence-transformers
langchain-huggingface
faiss-cpu
numpy
fastapi
uvicorn
sqlalchemy
python-dotenv
openai
anthropic
requests
neo4j
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
beautifulsoup4
duckduckgo-search
html2text